# BiBoGC/openapi-bibogc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BiBo's Grocery Store API",
    "description": "API cho hệ thống quản lý cửa hàng tạp hoá BiBo. Bao gồm quản lý sản phẩm, danh mục, nhà cung cấp, lô hàng và giao dịch kho.\n\n## Tính năng chính:\n- **Products**: Quản lý sản phẩm (CRUD)\n- **Categories**: Quản lý danh mục sản phẩm\n- **Suppliers**: Quản lý nhà cung cấp\n- **Product Batches**: Quản lý lô hàng (theo dõi hạn sử dụng)\n- **Stock Transactions**: Ghi nhận giao dịch kho (nhập/xuất/điều chỉnh)\n\n## Quy tắc chung:\n- Tất cả ID đều là GUID\n- Pagination sử dụng pageNumber (1-based) và pageSize\n- Soft delete: Dữ liệu bị xóa vẫn được lưu trong DB\n- Validation errors trả về 400 Bad Request với chi tiết lỗi",
    "version": "1.0.0",
    "contact": {
      "name": "BiBo's GC Support",
      "email": "support@bibogc.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5020",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/stocktransactions": {
      "get": {
        "tags": ["Stock Transactions"],
        "summary": "Lấy danh sách giao dịch kho (phân trang)",
        "description": "Lấy danh sách giao dịch kho với nhiều bộ lọc.\n\n**Lọc theo:**\n- productId: Lọc theo sản phẩm\n- supplierId: Lọc theo nhà cung cấp (xem đơn nhập của NCC cụ thể)\n- transactionType: Lọc theo loại giao dịch\n- fromDate/toDate: Lọc theo khoảng thời gian\n\n**Ví dụ:** Xem tất cả đơn nhập hàng từ NCC Coca-Cola:\n`GET /api/stocktransactions?supplierId=xxx&transactionType=1`",
        "operationId": "getStockTransactions",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Số trang",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Số lượng mỗi trang",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "searchTerm",
            "in": "query",
            "description": "Từ khóa tìm kiếm (tên SP, số lô, tên NCC)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "in": "query",
            "description": "Lọc theo ID sản phẩm",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "description": "Lọc theo ID nhà cung cấp (xem đơn nhập từ NCC cụ thể)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transactionType",
            "in": "query",
            "description": "Lọc theo loại giao dịch (1=Purchase, 2=Sale, 3=AdjustmentIn, 4=AdjustmentOut, 5=Damage, 6=Expiry, 7=Return, 8=SupplierReturn)",
            "schema": {
              "type": "integer",
              "enum": [1, 2, 3, 4, 5, 6, 7, 8]
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Lọc từ ngày (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Lọc đến ngày (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sắp xếp theo field",
            "schema": {
              "type": "string",
              "default": "TransactionDate"
            }
          },
          {
            "name": "sortDescending",
            "in": "query",
            "description": "Sắp xếp giảm dần (mới nhất trước)",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách giao dịch kho",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransactionResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Stock Transactions"],
        "summary": "Tạo giao dịch kho mới",
        "description": "Tạo một giao dịch kho mới và cập nhật tồn kho tương ứng.\n\n**Loại giao dịch:**\n- 1: Purchase - Nhập hàng từ NCC (bắt buộc supplierId)\n- 2: Sale - Bán hàng\n- 3: AdjustmentIn - Điều chỉnh tăng\n- 4: AdjustmentOut - Điều chỉnh giảm\n- 5: Damage - Hàng hư hỏng\n- 6: Expiry - Hàng hết hạn\n- 7: Return - Khách trả hàng\n- 8: SupplierReturn - Trả hàng NCC\n\n**Validation:**\n- productId: Bắt buộc, phải tồn tại\n- transactionType: Bắt buộc\n- quantity: Bắt buộc, > 0\n- unitPrice: >= 0\n- supplierId: Bắt buộc nếu type = Purchase",
        "operationId": "createStockTransaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockTransactionCommand"
              },
              "example": {
                "productId": "653eb0a5-fc22-41a8-9f2a-ee201b8524cf",
                "productBatchId": "075cc84e-9f38-400a-a74b-b79a83ee99c0",
                "supplierId": "cb143108-f662-4c89-b12d-e0b31fa48692",
                "transactionType": 1,
                "quantity": 100,
                "unitPrice": 10000,
                "notes": "Nhập hàng đợt 1 tháng 1/2026"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Giao dịch đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransactionDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy sản phẩm/lô hàng/NCC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocktransactions/{id}": {
      "get": {
        "tags": ["Stock Transactions"],
        "summary": "Lấy chi tiết giao dịch kho",
        "description": "Lấy thông tin chi tiết của một giao dịch kho.",
        "operationId": "getStockTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID giao dịch (GUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết giao dịch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransactionDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy giao dịch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocktransactions/purchase": {
      "post": {
        "tags": ["Stock Transactions"],
        "summary": "Tạo giao dịch nhập hàng (shortcut)",
        "description": "Shortcut để tạo giao dịch nhập hàng từ NCC.\nTương đương POST /api/stocktransactions với transactionType = 1",
        "operationId": "createPurchase",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseRequest"
              },
              "example": {
                "productId": "653eb0a5-fc22-41a8-9f2a-ee201b8524cf",
                "productBatchId": "075cc84e-9f38-400a-a74b-b79a83ee99c0",
                "supplierId": "cb143108-f662-4c89-b12d-e0b31fa48692",
                "quantity": 100,
                "unitPrice": 10000,
                "notes": "Nhập hàng từ Coca-Cola"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Giao dịch nhập hàng đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransactionDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocktransactions/sale": {
      "post": {
        "tags": ["Stock Transactions"],
        "summary": "Tạo giao dịch bán hàng (shortcut)",
        "description": "Shortcut để tạo giao dịch bán hàng.\nTương đương POST /api/stocktransactions với transactionType = 2",
        "operationId": "createSale",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleRequest"
              },
              "example": {
                "productId": "653eb0a5-fc22-41a8-9f2a-ee201b8524cf",
                "productBatchId": "075cc84e-9f38-400a-a74b-b79a83ee99c0",
                "quantity": 5,
                "unitPrice": 15000,
                "notes": "Bán lẻ"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Giao dịch bán hàng đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransactionDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation hoặc không đủ tồn kho",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocktransactions/adjustment": {
      "post": {
        "tags": ["Stock Transactions"],
        "summary": "Tạo giao dịch điều chỉnh kho (shortcut)",
        "description": "Shortcut để tạo giao dịch điều chỉnh tồn kho (tăng hoặc giảm).",
        "operationId": "createAdjustment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdjustmentRequest"
              },
              "example": {
                "productId": "653eb0a5-fc22-41a8-9f2a-ee201b8524cf",
                "productBatchId": "075cc84e-9f38-400a-a74b-b79a83ee99c0",
                "isIncrease": true,
                "quantity": 10,
                "unitPrice": 0,
                "notes": "Kiểm kê phát hiện thừa hàng"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Giao dịch điều chỉnh đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockTransactionDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StockTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID giao dịch"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm"
          },
          "productName": {
            "type": "string",
            "description": "Tên sản phẩm"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng"
          },
          "batchNumber": {
            "type": "string",
            "nullable": true,
            "description": "Số lô"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID nhà cung cấp"
          },
          "supplierName": {
            "type": "string",
            "nullable": true,
            "description": "Tên nhà cung cấp"
          },
          "sku": {
            "type": "string",
            "description": "Mã SKU"
          },
          "transactionType": {
            "type": "string",
            "description": "Loại giao dịch (Purchase, Sale, AdjustmentIn, AdjustmentOut, Damage, Expiry, Return, SupplierReturn)"
          },
          "quantity": {
            "type": "integer",
            "description": "Số lượng"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "description": "Đơn giá"
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal",
            "description": "Tổng tiền"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time",
            "description": "Ngày giao dịch"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Ghi chú"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true,
            "description": "Số tham chiếu"
          }
        }
      },
      "CreateStockTransactionCommand": {
        "type": "object",
        "required": ["productId", "transactionType", "quantity"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm (bắt buộc)"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng (tùy chọn)"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID nhà cung cấp (bắt buộc nếu Purchase)"
          },
          "transactionType": {
            "type": "integer",
            "enum": [1, 2, 3, 4, 5, 6, 7, 8],
            "description": "Loại giao dịch: 1=Purchase, 2=Sale, 3=AdjustmentIn, 4=AdjustmentOut, 5=Damage, 6=Expiry, 7=Return, 8=SupplierReturn"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng (bắt buộc, > 0)"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "minimum": 0,
            "default": 0,
            "description": "Đơn giá (>= 0)"
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "Ghi chú"
          }
        }
      },
      "CreatePurchaseRequest": {
        "type": "object",
        "required": ["productId", "supplierId", "quantity"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "description": "ID nhà cung cấp (bắt buộc)"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "description": "Đơn giá"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Ghi chú"
          }
        }
      },
      "CreateSaleRequest": {
        "type": "object",
        "required": ["productId", "quantity"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "description": "Đơn giá"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Ghi chú"
          }
        }
      },
      "CreateAdjustmentRequest": {
        "type": "object",
        "required": ["productId", "quantity", "isIncrease"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng"
          },
          "isIncrease": {
            "type": "boolean",
            "description": "true = tăng tồn, false = giảm tồn"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng điều chỉnh"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "default": 0,
            "description": "Đơn giá (thường = 0 cho điều chỉnh)"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Lý do điều chỉnh"
          }
        }
      },
...
      "PaginatedTransactionResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockTransactionDto"
            }
          },
          "pageNumber": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalCount": {
            "type": "integer"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}